feat: convert Integer backupIds to Long (#3328)

+++ qa/backup-restore-tests/src/test/java/io/camunda/tasklist/qa/backup/BackupRestoreTest.java
@@ -51,7 +51,7 @@ public class BackupRestoreTest {
   public static final String ZEEBE_INDEX_PREFIX = "backup-restore-test";
   public static final String VERSION = "SNAPSHOT";
   public static final String REPOSITORY_NAME = "testRepository";
-  public static final Integer BACKUP_ID = 123;
+  public static final Long BACKUP_ID = 123L;
   private static final Logger LOGGER = LoggerFactory.getLogger(BackupRestoreTest.class);
   private static final String TASKLIST_TEST_DOCKER_IMAGE = "camunda/tasklist";
